refactor: Package with initial new arquitecture

+++ src/Contracts/Gateways/MercadoPagoGateway.php
@@ -7,11 +7,11 @@ interface MercadoPagoGateway
     /**
      * @param array<string, mixed> $payload
      */
-    public function requestPayment(float $amount, string $description, array $payload = []): object;
+    public function requestPayment(int $amount, string $description, array $payload = []): object;
 
     public function getPaymentDetails(string $paymentId): object;
 
     public function getOrderDetails(string $orderId): object;
 
-    public function getRefund(string $paymentId, float $amount): object;
+    public function getRefund(string $paymentId, int $amount): object;
 }
